Suicide_Analysis_Project
In this project we tried to find out if we can predict suicide rates with maximum accuracy. We examined the subject in few parameters' we took in conclusion the average ages in different populations around the world. We also examined the ratio between male and female comparing to the suicide rates and so on.
In conclusion, we found out according to the models accuracy that we can almost predict suicide rates. (0.992 \ 0.999).